+++ src/main/java/com/qa/bookstore/student11/BookMapToJSONConverter.java
@@ -1,4 +1,4 @@
-package com.qa.bookstore.trainer;
+package com.qa.bookstore.student11;
 
 import java.util.Map;
